refactor: move cehckIsCompleted to an util function
